// networking/sylius-api-client/.ts.schemas.ts
export interface ProductVariantImageJsonldShopProductVariantRead {
  readonly "@context"?: ProductVariantImageJsonldShopProductVariantReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  readonly id?: number;
  path?: string;
}
// ...
export type ProductVariantJsonldShopProductVariantReadStorageType =
  typeof ProductVariantJsonldShopProductVariantReadStorageType[keyof typeof ProductVariantJsonldShopProductVariantReadStorageType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductVariantJsonldShopProductVariantReadStorageType = {
  WINE_CABINET: "WINE_CABINET",
  OTHER: "OTHER",
  AIR_CONDITIONED_UNDERGROUND_CELLAR: "AIR_CONDITIONED_UNDERGROUND_CELLAR",
  IDEALWINE_STORAGE_CELLAR: "IDEALWINE_STORAGE_CELLAR",
  CLASSIC_UNDERGROUND_CELLAR: "CLASSIC_UNDERGROUND_CELLAR",
  DEDICATED_RENTAL_STORAGE: "DEDICATED_RENTAL_STORAGE",
  SPECIALIZED_STORAGE: "SPECIALIZED_STORAGE",
  null: null,
} as const;

export type ProductVariantJsonldShopProductVariantReadProductVintage =
  ProductVintageJsonldShopProductVariantRead | null;

export type ProductVariantJsonldShopProductVariantReadPriceType =
  typeof ProductVariantJsonldShopProductVariantReadPriceType[keyof typeof ProductVariantJsonldShopProductVariantReadPriceType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductVariantJsonldShopProductVariantReadPriceType = {
  PRIX_SPECIAL: "PRIX_SPECIAL",
  PRIX_PROPRIETE: "PRIX_PROPRIETE",
  null: null,
} as const;

export type ProductVariantJsonldShopProductVariantReadPriceByCountry = {
  [key: string]: number;
} | null;

export type ProductVariantJsonldShopProductVariantReadOriginalPriceByCountry = {
  [key: string]: number;
} | null;

export type ProductVariantJsonldShopProductVariantReadOrigin =
  typeof ProductVariantJsonldShopProductVariantReadOrigin[keyof typeof ProductVariantJsonldShopProductVariantReadOrigin];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductVariantJsonldShopProductVariantReadOrigin = {
  OTHER: "OTHER",
  FAMILY_CELLAR: "FAMILY_CELLAR",
  WINE_MERCHANT: "WINE_MERCHANT",
  IDEALWINE_LOT: "IDEALWINE_LOT",
  STARRED_PROFESSIONAL: "STARRED_PROFESSIONAL",
  AUCTIONS: "AUCTIONS",
  ESTATE: "ESTATE",
  INDIVIDUAL: "INDIVIDUAL",
  PROFESSIONAL: "PROFESSIONAL",
  null: null,
} as const;

export type ProductVariantJsonldShopProductVariantReadFormat =
  typeof ProductVariantJsonldShopProductVariantReadFormat[keyof typeof ProductVariantJsonldShopProductVariantReadFormat];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductVariantJsonldShopProductVariantReadFormat = {
  BALTHAZAR: "BALTHAZAR",
  BOUTEILLE: "BOUTEILLE",
  CHOPINE: "CHOPINE",
  CLAVELIN: "CLAVELIN",
  DEMI_BOUTEILLE: "DEMI_BOUTEILLE",
  DOUBLE_MAGNUM: "DOUBLE_MAGNUM",
  FILLETTE: "FILLETTE",
  FLACON: "FLACON",
  FORMAT_50_CL: "FORMAT_50_CL",
  INCONNU: "INCONNU",
  IMPERIALE: "IMPERIALE",
  JEROBOAM: "JEROBOAM",
  LITRE: "LITRE",
  MAGNUM: "MAGNUM",
  MATHUSALEM: "MATHUSALEM",
  MELCHIOR: "MELCHIOR",
  NABUCHODONOSOR: "NABUCHODONOSOR",
  POT: "POT",
  REHOBOAM: "REHOBOAM",
  SALMANAZAR: "SALMANAZAR",
  null: null,
} as const;

export type ProductVariantJsonldShopProductVariantReadAppraisedBy =
  ExpertJsonldShopProductVariantRead | null;

export interface ProductVariantJsonldShopProductVariantRead {
  readonly "@context"?: ProductVariantJsonldShopProductVariantReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  additionalObservations?: string | null;
  alcoolProportion?: number | null;
  appraisedBy?: ProductVariantJsonldShopProductVariantReadAppraisedBy;
  auction?: boolean;
  averageEstimate?: number | null;
  code: string;
  CRD?: boolean;
  customVolume?: number | null;
  disgorgementYear?: number | null;
  format?: ProductVariantJsonldShopProductVariantReadFormat;
  hasOriginalWoodenCrate?: boolean;
  hasWineDexTag?: boolean;
  readonly id?: number;
  readonly inStock?: boolean;
  levels?: BottleLevelJsonldShopProductVariantRead[];
  limitedQuantityPerOrder?: number | null;
  lotDiscountBottleNumber?: number | null;
  lotDiscountPercentage?: number | null;
  name?: string | null;
  numberOfBottles?: number;
  observations?: BottleObservationJsonldShopProductVariantRead[];
  offlineSale?: boolean;
  onHand?: number;
  onHold?: number;
  optionValues?: string[];
  origin?: ProductVariantJsonldShopProductVariantReadOrigin;
  readonly originalPrice?: number;
  originalPriceByCountry?: ProductVariantJsonldShopProductVariantReadOriginalPriceByCountry;
  readonly price?: number;
  priceByCountry?: ProductVariantJsonldShopProductVariantReadPriceByCountry;
  priceType?: ProductVariantJsonldShopProductVariantReadPriceType;
  product?: ProductJsonldShopProductVariantRead;
  productVintage?: ProductVariantJsonldShopProductVariantReadProductVintage;
  publicProductVariantImages?: ProductVariantImageJsonldShopProductVariantRead[];
  quintessenceSale?: boolean;
  remark?: string | null;
  reserveBid?: number | null;
  restockable?: boolean;
  sagaRating?: number | null;
  secondHand?: boolean;
  seller?: CustomerJsonldShopProductVariantRead;
  storageType?: ProductVariantJsonldShopProductVariantReadStorageType;
  readonly VATRecoverable?: boolean;
  readonly volume?: number | null;
}
// ...
export type ProductVariantJsonldShopProductVariantReadContext =
  | string
  | ProductVariantJsonldShopProductVariantReadContextOneOf;
// ...
export interface ProductVariantShopSimilarProductVariantRead {
  auction?: boolean;
  code: string;
  readonly firstImage?: ProductVariantShopSimilarProductVariantReadFirstImage;
  format?: ProductVariantShopSimilarProductVariantReadFormat;
  hasLotDiscount?: boolean;
  readonly id?: number;
  lotDiscountBottleNumber?: number | null;
  lotDiscountPercentage?: number | null;
  numberOfBottles?: number;
  onHand?: number;
  onHold?: number;
  priceByCountry?: ProductVariantShopSimilarProductVariantReadPriceByCountry;
  product?: ProductShopSimilarProductVariantRead;
  productVintage?: ProductVariantShopSimilarProductVariantReadProductVintage;
}

export interface ProductVariantSimilarProductVariantsDTOShopSimilarProductVariantRead {
  sameAppellationVariants?: ProductVariantShopSimilarProductVariantRead[];
  sameOwnerVariants?: ProductVariantShopSimilarProductVariantRead[];
  sameRegionVariants?: ProductVariantShopSimilarProductVariantRead[];
}
// ...
export interface ProductJsonldShopProductVariantRead {
  readonly "@context"?: ProductJsonldShopProductVariantReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  alcoolProportion?: number | null;
  appellation?: string | null;
  readonly bio?: boolean;
  biologicProfile?: ProductJsonldShopProductVariantReadBiologicProfile;
  category?: ProductJsonldShopProductVariantReadCategory;
  classification?: ProductJsonldShopProductVariantReadClassification;
  color?: ProductJsonldShopProductVariantReadColor;
  description?: string | null;
  dominantAroma?: ProductJsonldShopProductVariantReadDominantAroma;
  estate?: ProductJsonldShopProductVariantReadEstate;
  images?: ProductImageJsonldShopProductVariantRead[];
  intensity?: ProductJsonldShopProductVariantReadIntensity;
  name?: string | null;
  readonly natural?: boolean;
  owner?: string | null;
  production?: number | null;
  productionUnit?: ProductJsonldShopProductVariantReadProductionUnit;
  profile?: ProductJsonldShopProductVariantReadProfile;
  region?: ProductJsonldShopProductVariantReadRegion;
  shortDescription?: string | null;
  surface?: number | null;
  tastingOccasion?: ProductJsonldShopProductVariantReadTastingOccasion;
  temperature?: number | null;
  readonly tripleA?: boolean;
  wineDishPairings?: string[];
  wineVarieties?: WineVarietyProductJsonldShopProductVariantRead[];
  yield?: number | null;
}
// ...
export type ExceptionCodeEnumJsonldId =
  typeof ExceptionCodeEnumJsonldId[keyof typeof ExceptionCodeEnumJsonldId];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ExceptionCodeEnumJsonldId = {
  a9a26902f: "a9a26902f",
  b5c3d807e: "b5c3d807e",
  kb4c6d5e7: "kb4c6d5e7",
  mxd8mt4ch: "mxd8mt4ch",
  r6p7b6jb2: "r6p7b6jb2",
  xhjDlAo4D: "xhjDlAo4D",
  eUnrn3LSx: "eUnrn3LSx",
  c2U7vrsq6: "c2U7vrsq6",
  K4m7L0sF6: "K4m7L0sF6",
  Oz1m7L0sF6: "Oz1m7L0sF6",
  O0gWKIx03w: "O0gWKIx03w",
} as const;
// ...
export interface CustomerJsonldShopProductVariantRead {
  readonly "@context"?: CustomerJsonldShopProductVariantReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  readonly id?: number;
}
// ...
export interface BottleObservationJsonldShopProductVariantRead {
  readonly "@context"?: BottleObservationJsonldShopProductVariantReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  readonly id?: number;
  quantity?: number;
  value?: BottleObservationJsonldShopProductVariantReadValue;
}
// ...
export interface BottleLevelJsonldShopProductVariantRead {
  readonly "@context"?: BottleLevelJsonldShopProductVariantReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  readonly id?: number;
  quantity?: number;
  value?: BottleLevelJsonldShopProductVariantReadValue;
}
// ...
export type AuctionWinningBidJsonldShopAuctionItemDtoReadShopUser =
  ShopUserJsonldShopAuctionItemDtoRead | null;

export interface AuctionWinningBidJsonldShopAuctionItemDtoRead {
  readonly "@context"?: AuctionWinningBidJsonldShopAuctionItemDtoReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  auctionItem?: ProductVariantJsonldShopAuctionItemDtoRead;
  createdAt?: string;
  price?: number;
  shopUser?: AuctionWinningBidJsonldShopAuctionItemDtoReadShopUser;
}
// ...
export type AuctionItemDTOJsonldShopAuctionItemDtoReadVariantInAuctionCatalog =
  ProductVariantInAuctionCatalogJsonldShopAuctionItemDtoRead | null;

export type AuctionItemDTOJsonldShopAuctionItemDtoReadCurrentWinner =
  ShopUserJsonldShopAuctionItemDtoRead | null;

export type AuctionItemDTOJsonldShopAuctionItemDtoReadAuctionCatalog =
  AuctionCatalogJsonldShopAuctionItemDtoRead | null;

export interface AuctionItemDTOJsonldShopAuctionItemDtoRead {
  readonly "@context"?: AuctionItemDTOJsonldShopAuctionItemDtoReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  auctionCatalog: AuctionItemDTOJsonldShopAuctionItemDtoReadAuctionCatalog;
  buyerFeePercentage: number;
  code: string;
  currentWinner?: AuctionItemDTOJsonldShopAuctionItemDtoReadCurrentWinner;
  endDate?: string;
  followersCount: number;
  highestBid?: number | null;
  historicWinningBids: AuctionWinningBidJsonldShopAuctionItemDtoRead[];
  initialEndDate?: string;
  numberOfBids: number;
  overtimeBiddingEnabled?: boolean;
  reserveBid: number;
  readonly serverTimeMilliseconds?: number;
  thresholdBid?: number | null;
  variantInAuctionCatalog: AuctionItemDTOJsonldShopAuctionItemDtoReadVariantInAuctionCatalog;
  VATPercentage: number;
}
// ...
export type AuctionItemDTOJsonldShopAuctionItemDtoReadContextOneOf = {
  "@vocab": string;
  hydra: AuctionItemDTOJsonldShopAuctionItemDtoReadContextOneOfHydra;
  [key: string]: any;
};

export type AuctionItemDTOJsonldShopAuctionItemDtoReadContext =
  | string
  | AuctionItemDTOJsonldShopAuctionItemDtoReadContextOneOf;

export interface AuctionItemDTOAuctionBidResultDTOShopAuctionItemDtoRead {
  bidId?: number | null;
  result?: string;
}
// ...
export type AuctionItemDTOAuctionBidDTOJsonldContextOneOf = {
  "@vocab": string;
  hydra: AuctionItemDTOAuctionBidDTOJsonldContextOneOfHydra;
  [key: string]: any;
};

export type AuctionItemDTOAuctionBidDTOJsonldContext =
  | string
  | AuctionItemDTOAuctionBidDTOJsonldContextOneOf;

export interface AuctionItemDTOAuctionBidDTOJsonld {
  readonly "@context"?: AuctionItemDTOAuctionBidDTOJsonldContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  amount?: number;
  threshold?: boolean;
}
// ...
export interface AuctionCatalogJsonldShopAuctionItemDtoRead {
  readonly "@context"?: AuctionCatalogJsonldShopAuctionItemDtoReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  endDate?: string | null;
  readonly id?: number;
  name?: string;
  startDate?: string | null;
}
// ...
export type AuctionBidResultEnumJsonldId =
  typeof AuctionBidResultEnumJsonldId[keyof typeof AuctionBidResultEnumJsonldId];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AuctionBidResultEnumJsonldId = {
  WINNING: "WINNING",
  SURPASSED_BY_THRESHOLD: "SURPASSED_BY_THRESHOLD",
  OVERTAKING: "OVERTAKING",
} as const;